Update the UI for the insert/delete completion functionality
Refactor Routine to Habit along the way.

The app currently works well, but the habit streak functionality is not working. It always displays zeros in the current/latest streak cards. Streak functionality is also required to determine the color of an ambiguous date in RoutineCalendar, such as one that is associated with NotDue or Skipped habit status. So these days are displayed with the green (completed) color at the moment.

The is also a lot of minor improvements and fixes to be done including refactoring Routine to Habit in the scope of the whole project