Drop support for period separation
Each period is independent now. The actual functionality is not deleted, but the  periodSeparationEnabled property always has the value of true.
The reason for dropping the support is because it leads to unexpected results for the user and is also tricky to maintain. Not sure whether it's essential at the current state of the app.

Also limit the number of due days and the number of rest days in AlternateDaysSchedule to 99 each. That's because it takes long to compute habit statuses for such big periods, while the use cases for that are not clear.