Load data for each period independently and cash streaks
Getting the entire list of completions for an old habit (e.g. started in 2010) takes way too much time. This commit solves this problem by fetching only data for the requested period + a schedule period margin.

Streaks from past periods are now computed only once and cashed in the database. Streaks for the current period are not cashed because they can be frequently changed.
This approach introduces the need to update cashed streaks for the corresponding period whenever a completion is inserted. So I also added the logic for that.

Additionally, made some code quality improvements.