Fix unexpected behavior
When the first period date was not completed and backlog was disabled, that date was not included in the streak, which is undesired. So dropped support the following feature: "not due dates before the completed date are not included in the streak if backlog is disabled" because it results in unexpected behavior.